Move cpp_ui under cpp_audio and update build
