+++ CMakeLists.txt
@@ -2,4 +2,3 @@ cmake_minimum_required(VERSION 3.14)
 project(DIY_AV_CPP)
 
 add_subdirectory(src/cpp_audio)
-add_subdirectory(src/cpp_ui)
