+++ src/cpp_audio/ui/NoiseGeneratorDialog.cpp
@@ -1,5 +1,5 @@
-#include "../cpp_audio/Track.h"
-#include "../cpp_audio/SynthFunctions.h"
+#include "../Track.h"
+#include "../SynthFunctions.h"
 #include <juce_audio_formats/juce_audio_formats.h>
 #include <juce_audio_utils/juce_audio_utils.h>
 #include <juce_gui_basics/juce_gui_basics.h>
