+++ src/cpp_audio/ui/NoiseGeneratorDialog.h
