+++ src/cpp_audio/ui/OverlayClipDialog.cpp
