+++ src/cpp_audio/ui/OverlayClipDialog.h
