+++ src/cpp_audio/ui/OverlayClipPanel.cpp
