+++ src/cpp_audio/ui/OverlayClipPanel.h
