+++ src/cpp_audio/ui/StepConfigPanel.cpp
