+++ src/cpp_audio/ui/StepPreviewComponent.cpp
